Consolidate 8 script files into scripts/ci.sh and scripts/ci.py
8 single-purpose files (5 shell, 3 Python) for what's really one CI
toolkit was adding up to more file-count overhead than the logic
warranted. Merged into two files, one per language, each with a
function/subcommand per responsibility instead of one file per
responsibility:

scripts/ci.sh (function dispatch via `"$@"` at the bottom, called as
`bash scripts/ci.sh <function> [args...]`):
  fetch_eval_harness, fingerprint_main_content, check_cache_hit,
  author_and_evaluate, author_and_evaluate_baseline

scripts/ci.py (argparse subcommands, called as
`python3 scripts/ci.py <subcommand> ...`):
  fingerprint, bundle write, bundle verify, print-outputs

The two files still call each other exactly as before (ci.sh's functions
shell out to `python3 scripts/ci.py ...`) -- this is a rename/regroup, not
a behavior change. ci.sh deliberately has no global `set -e`:
author_and_evaluate needs to tolerate a failing author-app.sh (so the PR
comment still shows something), while fetch_eval_harness,
fingerprint_main_content, and author_and_evaluate_baseline each wrap their
own body in a `(set -euo pipefail; ...)` subshell for the strictness they
individually need.

Verified byte-for-byte identical behavior against the previous 8 files:
same fingerprint determinism/invalidation suite, same symlink-safety
tests, same bundle write/verify round-trip and rejection cases, same
print-outputs results against the 3 real e2e-run metrics.json files (and
their edge cases), same check_cache_hit and author_and_evaluate
cache-hit short-circuit behavior. Both workflow YAML files re-validate
against the EAS schema and stay under the 16 KiB limit
(16,298 / 9,087 bytes).
