+++ .eas/workflows/skill-eval-ci.yml
@@ -1,8 +1,7 @@
 # Skill eval for PRs touching plugins/expo/skills/**. Advisory, never blocks
 # merge. Authors+evaluates 3 PRDs vs main and this PR. "_main" jobs cache via
 # fingerprint+restore_cache; only skill-eval-main-baseline.yml writes that
-# cache. Logic in scripts/*: EAS caps files at 16 KiB, see
-# plugins/expo/skills/eas-workflows/SKILL.md.
+# cache. Logic in scripts/*: EAS caps workflow files at 16 KiB.
 name: skill-eval-ci
 
 on:
@@ -39,7 +38,7 @@ jobs:
       - name: fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh extract > fp-notes-main.json
+          bash scripts/ci.sh fingerprint_main_content extract > fp-notes-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fp-notes-main.json'))['fingerprint'])")"
 
       - name: restore cache
@@ -50,12 +49,12 @@ jobs:
 
       - name: check cache
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-notes-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-notes-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-notes-main /tmp/plugins-main/plugins/expo skill-eval-report-notes-main "${SCENARIO}" skill-eval-report-notes-main.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-notes-main /tmp/plugins-main/plugins/expo skill-eval-report-notes-main "${SCENARIO}" skill-eval-report-notes-main.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
@@ -99,7 +98,7 @@ jobs:
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh false - "$(pwd)/plugins/expo" skill-eval-report-notes-pr "${SCENARIO}" skill-eval-report-notes-pr.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate false - "$(pwd)/plugins/expo" skill-eval-report-notes-pr "${SCENARIO}" skill-eval-report-notes-pr.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
@@ -144,7 +143,7 @@ jobs:
       - name: fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh extract > fp-hot_chocolate-main.json
+          bash scripts/ci.sh fingerprint_main_content extract > fp-hot_chocolate-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fp-hot_chocolate-main.json'))['fingerprint'])")"
 
       - name: restore cache
@@ -155,12 +154,12 @@ jobs:
 
       - name: check cache
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-hot_chocolate-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-hot_chocolate-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-hot_chocolate-main /tmp/plugins-main/plugins/expo skill-eval-report-hot_chocolate-main "${SCENARIO}" skill-eval-report-hot_chocolate-main.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-hot_chocolate-main /tmp/plugins-main/plugins/expo skill-eval-report-hot_chocolate-main "${SCENARIO}" skill-eval-report-hot_chocolate-main.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
@@ -204,7 +203,7 @@ jobs:
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh false - "$(pwd)/plugins/expo" skill-eval-report-hot_chocolate-pr "${SCENARIO}" skill-eval-report-hot_chocolate-pr.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate false - "$(pwd)/plugins/expo" skill-eval-report-hot_chocolate-pr "${SCENARIO}" skill-eval-report-hot_chocolate-pr.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
@@ -249,7 +248,7 @@ jobs:
       - name: fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh extract > fp-wiki_reader-main.json
+          bash scripts/ci.sh fingerprint_main_content extract > fp-wiki_reader-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fp-wiki_reader-main.json'))['fingerprint'])")"
 
       - name: restore cache
@@ -260,12 +259,12 @@ jobs:
 
       - name: check cache
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-wiki_reader-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-wiki_reader-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-wiki_reader-main /tmp/plugins-main/plugins/expo skill-eval-report-wiki_reader-main "${SCENARIO}" skill-eval-report-wiki_reader-main.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate "${{ steps.check_cache.outputs.cache_hit }}" cached-baseline-wiki_reader-main /tmp/plugins-main/plugins/expo skill-eval-report-wiki_reader-main "${SCENARIO}" skill-eval-report-wiki_reader-main.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
@@ -309,7 +308,7 @@ jobs:
       - name: author and evaluate
         id: run
         run: |
-          eval "$(bash scripts/author-and-evaluate.sh false - "$(pwd)/plugins/expo" skill-eval-report-wiki_reader-pr "${SCENARIO}" skill-eval-report-wiki_reader-pr.tar.gz")"
+          eval "$(bash scripts/ci.sh author_and_evaluate false - "$(pwd)/plugins/expo" skill-eval-report-wiki_reader-pr "${SCENARIO}" skill-eval-report-wiki_reader-pr.tar.gz")"
           set-output report "$report"
           set-output expected "$expected"
           set-output detected "$detected"
