+++ .eas/workflows/skill-eval-main-baseline.yml
@@ -47,7 +47,7 @@ jobs:
       - name: compute main fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh checkout > fingerprint-notes-main.json
+          bash scripts/ci.sh fingerprint_main_content checkout > fingerprint-notes-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fingerprint-notes-main.json'))['fingerprint'])")"
 
       - name: restore cached main baseline
@@ -58,11 +58,11 @@ jobs:
 
       - name: check cache hit
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-notes-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-notes-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate if needed
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
-        run: bash scripts/author-and-evaluate-baseline.sh "$(pwd)/plugins/expo" cached-baseline-notes-main skills_available_unmentioned
+        run: bash scripts/ci.sh author_and_evaluate_baseline "$(pwd)/plugins/expo" cached-baseline-notes-main skills_available_unmentioned
 
       # should_save folds the cache-miss check and the trusted-branch check
       # into one bash-computed output, so save_cache only needs a single
@@ -71,7 +71,7 @@ jobs:
         id: manifest
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
         run: |
-          python3 scripts/validate-result-bundle.py write \
+          python3 scripts/ci.py bundle write \
             --bundle-dir cached-baseline-notes-main \
             --fingerprint-json fingerprint-notes-main.json \
             --status success --kind baseline \
@@ -114,7 +114,7 @@ jobs:
       - name: compute main fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh checkout > fingerprint-hot_chocolate-main.json
+          bash scripts/ci.sh fingerprint_main_content checkout > fingerprint-hot_chocolate-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fingerprint-hot_chocolate-main.json'))['fingerprint'])")"
 
       - name: restore cached main baseline
@@ -125,17 +125,17 @@ jobs:
 
       - name: check cache hit
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-hot_chocolate-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-hot_chocolate-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate if needed
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
-        run: bash scripts/author-and-evaluate-baseline.sh "$(pwd)/plugins/expo" cached-baseline-hot_chocolate-main skills_available_unmentioned
+        run: bash scripts/ci.sh author_and_evaluate_baseline "$(pwd)/plugins/expo" cached-baseline-hot_chocolate-main skills_available_unmentioned
 
       - name: write baseline manifest
         id: manifest
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
         run: |
-          python3 scripts/validate-result-bundle.py write \
+          python3 scripts/ci.py bundle write \
             --bundle-dir cached-baseline-hot_chocolate-main \
             --fingerprint-json fingerprint-hot_chocolate-main.json \
             --status success --kind baseline \
@@ -178,7 +178,7 @@ jobs:
       - name: compute main fingerprint
         id: fingerprint
         run: |
-          bash scripts/fingerprint-main-content.sh checkout > fingerprint-wiki_reader-main.json
+          bash scripts/ci.sh fingerprint_main_content checkout > fingerprint-wiki_reader-main.json
           set-output fingerprint "$(python3 -c "import json; print(json.load(open('fingerprint-wiki_reader-main.json'))['fingerprint'])")"
 
       - name: restore cached main baseline
@@ -189,17 +189,17 @@ jobs:
 
       - name: check cache hit
         id: check_cache
-        run: set-output cache_hit "$(bash scripts/check-cache-hit.sh cached-baseline-wiki_reader-main "${{ steps.fingerprint.outputs.fingerprint }}")"
+        run: set-output cache_hit "$(bash scripts/ci.sh check_cache_hit cached-baseline-wiki_reader-main "${{ steps.fingerprint.outputs.fingerprint }}")"
 
       - name: author and evaluate if needed
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
-        run: bash scripts/author-and-evaluate-baseline.sh "$(pwd)/plugins/expo" cached-baseline-wiki_reader-main skills_available_unmentioned
+        run: bash scripts/ci.sh author_and_evaluate_baseline "$(pwd)/plugins/expo" cached-baseline-wiki_reader-main skills_available_unmentioned
 
       - name: write baseline manifest
         id: manifest
         if: ${{ steps.check_cache.outputs.cache_hit == 'false' }}
         run: |
-          python3 scripts/validate-result-bundle.py write \
+          python3 scripts/ci.py bundle write \
             --bundle-dir cached-baseline-wiki_reader-main \
             --fingerprint-json fingerprint-wiki_reader-main.json \
             --status success --kind baseline \
